project/requirements.txt
# Kafka
kafka-python>=2.3.0

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.0

# Web API
flask==3.0.2

# ML
scikit-learn==1.4.0
joblib==1.3.2

# Utilities
python-dotenv==1.0.1
